port pmap to support 4-level page mapping and modify other related to make the code compiled

+++ kern/mem.h
@@ -48,8 +48,8 @@ typedef struct pageinfo {
 	struct pageinfo	*free_next;	// Next page number on free list
 	int32_t	refcount;		// Reference count on allocated pages
 #if LAB >= 5
-	uint32_t home;			// Remote reference to page's home
-	uint32_t shared;		// Other nodes I've given RRs to
+	intptr_t home;			// Remote reference to page's home
+	intptr_t shared;		// Other nodes I've given RRs to
 	struct pageinfo *homelist;	// My pages with homes at this physaddr
 	struct pageinfo *homenext;	// Next pointer on homelist
 #endif
